requirements.txt
fastapi
requests
beautifulsoup4
imdbpy
uvicorn
